requirements.txt
anyio==3.7.0
boto3==1.28.10
botocore==1.31.11
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
mangum==0.17.0
ptvsd==4.3.2
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
